Added more explicit error message when a trying to find the vertex buffer for a specific attribute
The compiler might strip away the vertex attribute if it has not been used in the vert / frag shader pipe. This creates a problem when trying to get the buffer for that attribute and bind it to a vao. This message notifies the user but a better solution is needed. Preferably the removal of the attribute in mAttributes  after all the buffer have been created