Use the path mapping system for findProjectFilePath
Project files are now located using the path mapping system, as they
should be. The unique case of locating the project.json (which can't use
the path mapping system as it hasn't been initialized yet - the chicken
or the egg) is separated out. The project.json searching will be
simplified in future when the directory structures in the first two
contexts are simplified. This also resolves service config loading in
Napkin, plus stops config.json files next to the binary being
preferenced over the one pointed to in the project.json.